fix(R4): review — honest int8 bound, concurrent bias coverage, hardcoded-value guards
All findings from the R4 adversarial review applied:

- MEDIUM: the documented MPNet int8 figure (cosine ≥ 0.9989) was
  measured on 2 texts; over the full golden the worst case is 0.997874
  (the punctuation text). Docs corrected to ≥ 0.9978, and the golden
  matrix now ENFORCES it: an int8MinCos column loads WithInt8 and
  validates cosine per case, so every documented int8 figure is
  test-backed instead of asserted in three documents and enforced in
  zero tests.
- MEDIUM: TestGoldenMatrix going serial left the per-head biasDelta
  reads with no concurrent coverage (the test the comment cited for
  worker-invariance loads a BERT model that has no such path). New
  TestGoldenMPNetParallelMatchesSerial: a ≥300-token input through
  default-worker and serial loads must be BIT-identical — covering the
  shared bias table read from inside Pool.Run goroutines, and the far
  buckets concurrently.
- LOW: golden cases previously topped out at 91 tokens, so the
  log-spaced far buckets and the max-distance clamp were only unit-
  tested. convert.py grows a ~400-token case (324 tokens for MPNet);
  the regenerated 12-case golden pins them end to end. fp32 stays at
  maxAbs 3.3e-7.
- LOW: the scale-then-add at the bias site is a fusable x*y+z, and Go
  emits an FMA on arm64 — so 'rounds the same as HF' was not literally
  true there. An explicit float32() conversion forces the product to
  round; architectures are now bit-identical (MPNet golden + parallel
  test re-proven under qemu-aarch64).
- LOW: HF's MPNet code HARDCODES num_buckets=32 (compute_position_bias
  never reads the config value) and padding_idx=1. A checkpoint
  declaring anything else would silently diverge; validate() and
  convert.py now refuse both, with derive tests for each refusal.

Race suite green in ~36s; arm64 model tests and the MPNet golden pass
under qemu.

+++ internal/model/bert.go
@@ -568,13 +568,16 @@ func (m *Model) encodeWorkers(ids []int64, workers int) (*scratch, error) {
 			tensor.MatMulSerial(sc, qhH, khH, seq, dh, seq)
 			if m.relBias != nil {
 				// MPNet: scores/√dh + bias(j−i), matching HF's order
-				// (scale first, then add) so accumulation rounds the same.
+				// (scale first, then add). The explicit float32 conversion
+				// forces the product to round before the add — without it
+				// the compiler fuses an FMA on arm64, and results stop
+				// being bit-identical across architectures.
 				bd := s.biasDelta[h*(2*seq-1):]
 				for i := range seq {
 					row := sc[i*seq : i*seq+seq]
 					bdi := bd[seq-1-i:] // index j-i+seq-1 = (seq-1-i)+j
 					for j := range row {
-						row[j] = row[j]*scale + bdi[j]
+						row[j] = float32(row[j]*scale) + bdi[j]
 					}
 				}
 			} else {
